refactor(PlacementConstraintsPartial): extract partial
Closes DCOS-19557
